Use vals in VirtualMachineFactory
We don't need to create a separate instance of the ByteCodeFactory or
ByteCodeParser each time the factory functions are called (indeed, we
can't create a new instance of ByteCodeFactoryImpl, because it's a
singleton), so it makes sense to use a val rather than a def. In the
case of ByteCodeParserImpl, which isn't a singleton, but which we create
based on the default ByteCodeFactory instance, a lazy val, which defers
creation until it is needed, is appropriate.